use core Modal for export - which fixes mobile view
