--- projects/packages/forms/src/dashboard/inbox/export-modal/index.js
@@ -1,85 +1,0 @@
-/**
- * External dependencies
- */
-import { Gridicon, JetpackFooter } from '@automattic/jetpack-components';
-import { useSelect } from '@wordpress/data';
-import { useCallback, useEffect, useRef } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
-/**
- * Internal dependencies
- */
-import { config } from '../../';
-import { STORE_NAME } from '../../state';
-import CSVExport from './csv';
-import GoogleDriveExport from './google-drive';
-
-import './style.scss';
-
-const ExportModal = ( { isVisible, onClose } ) => {
-	const backdrop = useRef();
-	const { selected, currentQuery } = useSelect( select => {
-		const { getSelectedResponsesFromCurrentDataset, getCurrentQuery } = select( STORE_NAME );
-		return { selected: getSelectedResponsesFromCurrentDataset(), currentQuery: getCurrentQuery() };
-	}, [] );
-	useEffect( () => {
-		backdrop.current?.addEventListener( 'click', event => {
-			if ( event.target !== backdrop.current ) {
-				return;
-			}
-
-			onClose();
-		} );
-	}, [ isVisible, onClose ] );
-
-	const exportResponses = useCallback(
-		( action, nonceName ) => {
-			const data = new FormData();
-			data.append( 'action', action );
-			data.append( nonceName, config( 'exportNonce' ) );
-			selected.forEach( id => data.append( 'selected[]', id ) );
-			data.append( 'post', currentQuery.parent || 'all' );
-			data.append( 'search', currentQuery.search || '' );
-			data.append( 'status', currentQuery.status );
-			if ( currentQuery.before && currentQuery.after ) {
-				data.append( 'before', currentQuery.before );
-				data.append( 'after', currentQuery.after );
-			}
-			return fetch( window.ajaxurl, { method: 'POST', body: data } );
-		},
-		[ currentQuery, selected ]
-	);
-
-	if ( ! isVisible ) {
-		return null;
-	}
-
-	return (
-		<div ref={ backdrop } className="jp-forms__export-modal">
-			<div className="jp-forms__export-modal-wrapper">
-				<button className="jp-forms__export-modal-close-button" onClick={ onClose }>
-					<Gridicon icon="cross" size={ 18 } />
-				</button>
-
-				<div className="jp-forms__export-modal-header">
-					<h1 className="jp-forms__export-modal-header-title">
-						{ __( 'Export your Form Responses', 'jetpack-forms' ) }
-					</h1>
-					<p className="jp-forms__export-modal-header-subtitle">
-						{ __( 'Choose your favorite file format or export destination:', 'jetpack-forms' ) }
-					</p>
-				</div>
-				<div className="jp-forms__export-modal-content">
-					<CSVExport onExport={ exportResponses } />
-					<GoogleDriveExport onExport={ exportResponses } />
-				</div>
-
-				<JetpackFooter
-					className="jp-forms__export-modal-footer"
-					moduleName={ __( 'Jetpack Forms', 'jetpack-forms' ) }
-				/>
-			</div>
-		</div>
-	);
-};
-
-export default ExportModal;
