+++ projects/packages/forms/src/dashboard/inbox/export-responses/csv.js
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import { Button } from '@wordpress/components';
 import { useCallback } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import clsx from 'clsx';
@@ -58,9 +59,9 @@ const CSVExport = ( { onExport } ) => {
 					{ __( 'Download your form response data as a CSV file.', 'jetpack-forms' ) }
 				</div>
 				<div className="jp-forms__export-modal-card-body-cta">
-					<button className={ buttonClasses } onClick={ downloadCSV }>
+					<Button variant="primary" className={ buttonClasses } onClick={ downloadCSV }>
 						{ __( 'Download', 'jetpack-forms' ) }
-					</button>
+					</Button>
 				</div>
 			</div>
 		</div>
