+++ projects/packages/forms/src/dashboard/inbox/export-responses/style.scss
@@ -1,56 +1,3 @@
-.jp-forms__export-modal {
-	align-items: center;
-	background: rgba(0, 0, 0, 0.3);
-	box-sizing: border-box;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	padding: 0;
-	position: fixed;
-	top: 0;
-	left: 0;
-	bottom: 0;
-	right: 0;
-	z-index: 9999;
-
-	@media (min-width: 601px) {
-		padding: 40px 60px;
-	}
-}
-
-.jp-forms__export-modal-wrapper {
-	background-color: #fff;
-	display: flex;
-	flex-direction: column;
-	max-width: 680px;
-	padding: 24px 48px 0;
-	position: relative;
-}
-
-.jp-forms__export-modal-close-button {
-	background: transparent;
-	border: 0;
-	cursor: pointer;
-	height: 24px;
-	padding: 3px;
-	position: absolute;
-	right: 8px;
-	top: 8px;
-	width: 24px;
-}
-
-.jp-forms__export-modal-header {
-	margin-bottom: 32px;
-}
-
-.jp-forms__export-modal-header-title {
-	font-style: normal;
-	font-weight: 700;
-	font-size: 24px;
-	line-height: 32px;
-	margin-bottom: 8px;
-}
-
 .jp-forms__export-modal-header-subtitle {
 	font-style: normal;
 	font-weight: 400;
@@ -59,28 +6,14 @@
 	letter-spacing: -0.02em;
 	color: #000000;
 	padding: 0;
-	margin-top: 8px;
 	margin-bottom: 0;
-}
-
-.jp-forms__export-modal-content {
-	display: flex;
-	flex-grow: 1;
-	flex-direction: column;
-	padding: 0px;
-	gap: 32px;
 }
 
 .jp-forms__export-modal-footer {
 	box-sizing: border-box;
 	display: flex;
 	justify-content: space-between;
-	padding: 48px 16px 24px;
-}
-
-.jp-forms__export-modal-footer-column {
-	display: flex;
-	align-items: center;
+	padding: 0 16px;
 }
 
 /* override JP logo */
@@ -162,16 +95,6 @@
 
 .jp-forms__export-modal-card-body-cta {
 	align-self: flex-end;
-}
-
-.jp-forms__export-modal-card-body-cta .button-primary.export-button {
-	font-style: normal;
-	font-weight: 400;
-	letter-spacing: -0.01em;
-	color: #FFFFFF;
-	background: #000000;
-	border: 1px solid #FFFFFF;
-	border-radius: 4px;
 }
 
 .jp-forms__export-modal-card-beta-badge {
